app/requirements.txt
# requirements.txt
# Core dependencies
customtkinter>=5.2.0
opencv-python>=4.5.0
numpy>=1.23.0
pillow>=9.0.0
matplotlib>=3.5.0
psutil>=5.9.0
requests>=2.28.0
schedule>=1.1.0
packaging>=21.3
setuptools<81  # Pinned to suppress pygame pkg_resources warning

# Optional dependencies
tkinterdnd2>=0.3.0  # For drag-and-drop support
speedtest-cli>=2.1.3  # For network speed testing
google-api-python-client>=2.80.0  # For YouTube upload
google-auth-oauthlib>=0.4.6  # For YouTube authentication
pygame>=2.1.0  # For music preview
torch>=2.0.0  # For ML-based video type detection

# Testing dependencies
pytest>=7.0.0

# Note: FFmpeg (>4.0) must be installed separately and available in system PATH